// frontend/src/types/messstelle/FahrzeugOptions.ts
export default interface FahrzeugOptions {
  kraftfahrzeugverkehr: boolean;
  schwerverkehr: boolean;
  gueterverkehr: boolean;
  schwerverkehrsanteilProzent: boolean;
  gueterverkehrsanteilProzent: boolean;
  radverkehr: boolean;
  fussverkehr: boolean;
  lastkraftwagen: boolean;
  lastzuege: boolean;
  busse: boolean;
  kraftraeder: boolean;
  personenkraftwagen: boolean;
  lieferwagen: boolean;
}